fixed migration error

+++ database/migrations/2019_10_13_134350_create_education_history.php
@@ -22,7 +22,7 @@ class CreateEducationHistory extends Migration
             $table->string('final_score');
             $table->string('institution');
             $table->string('description');
-            $table->integer('candidate_id')->unsigned();
+            $table->bigInteger('candidate_id')->unsigned();
             $table->foreign('candidate_id')->references('id')->on('candidates')->onDelete('cascade');
             $table->timestamps();
         });
